+++ src/com/excilys/formation/computerdatabase/persistence/Crudable.java
@@ -13,12 +13,12 @@ public interface Crudable<T> extends Findable<T> {
 	 * @param toCreate the new object to persist
 	 * @return the generated if
 	 */
-	int create(T toCreate);
+	long create(T toCreate);
 	
 	/**
 	 * 
 	 * @param id the id of the row to delete
 	 * @return the number of affected rows
 	 */
-	int delete(int id);
+	int delete(long id);
 }
